requirements.txt
amqp==2.3.2
appdirs==1.4.3
asgi-redis==1.4.3
asgiref==2.3.2
async-timeout==2.0.1
bandit==1.5.1
billiard==3.5.0.4
boto==2.49.0
boto3==1.9.16
botocore==1.12.16
celery==4.1.1
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
coverage==4.3.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-appconf==1.0.2
django-betterforms==1.2
django-celery-email==2.0.1
django-classy-tags==0.8.0
django-colorfield==0.1.15
django-contrib-comments==1.9.0
django-cors-headers==2.4.0
django-crum==0.7.3
django-debug-toolbar==1.10.1
django-extra-views==0.11.0
django-fcm==0.1.1
django-filter==2.0.0
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-postgrespool==0.3.1
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-storages==1.7.1
django-taggit==0.23.0
django-taggit-templatetags2==1.6.1
django-versatileimagefield==1.10
django-widget-tweaks==1.4.3
djangorestframework==3.8.2
docutils==0.14
drfdocs==0.0.11
flake8==3.5.0
flake8-polyfill==1.0.2
gevent==1.3.6
gitdb2==2.0.4
GitPython==2.1.11
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonfield==2.0.2
kombu==4.2.1
mando==0.6.4
MarkupSafe==1.0
mccabe==0.6.1
msgpack-python==0.5.6
oauthlib==2.1.0
olefile==0.46
openapi-codec==1.3.2
pbr==4.3.0
Pillow==4.0.0
psycopg2==2.7.5
pycodestyle==2.4.0
pyflakes==2.0.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
radon==2.3.1
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
simplejson==3.16.0
six==1.11.0
smmap2==2.0.4
SQLAlchemy==1.2.12
sqlparse==0.2.4
stevedore==1.29.0
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
